+++ test/smokes/metrics_fileinfo/unicode_filepath/master-test/🗄️.js
@@ -1,0 +1,2 @@
+const test = 'adsf';
+test = 'adsf';