+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Updated tools:
 - **ESLint** 7.30.0 -> 7.31.0 [#2544](https://github.com/sider/runners/pull/2544)
 - **Goodcheck** 3.0.3 -> 3.1.0 [#2542](https://github.com/sider/runners/pull/2542)
 
+Misc:
+
+- **Metrics File Info** Fix unicode filepath [#2549](https://github.com/sider/runners/pull/2549)
+
 ## 0.52.1
 
 [Full diff](https://github.com/sider/runners/compare/0.52.0...0.52.1)